Policies created

+++ app/Http/Controllers/HobbyController.php
@@ -7,6 +7,7 @@ use App\Tag;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Session;
 use Intervention\Image\Facades\Image;
+use Illuminate\Support\Facades\Gate;
 
 class HobbyController extends Controller
 {
@@ -99,6 +100,8 @@ class HobbyController extends Controller
      */
     public function edit(Hobby $hobby)
     {
+        abort_unless(Gate::allows('update' , $hobby), 403,'You are not authorized.');
+
         return view('hobby.edit')->with([
           'hobby'=>$hobby,
             'message_success' => Session::get('message_success'),
@@ -115,6 +118,8 @@ class HobbyController extends Controller
      */
     public function update(Request $request, Hobby $hobby)
     {
+        abort_unless(Gate::allows('update' , $hobby), 403,'You are not authorized.');
+
       $request->validate([
         'name' => 'required|min:3',
         'description' => 'required|min:5',
@@ -143,6 +148,8 @@ class HobbyController extends Controller
      */
     public function destroy(Hobby $hobby)
     {
+        abort_unless(Gate::allows('delete' , $hobby), 403,'You are not authorized.');
+
         $oldName = $hobby->name;
         $hobby->delete();
         return $this->index()->with(
